+++ Stonks/src/portfolio/model/IPortfolios.java
@@ -57,7 +57,7 @@ public interface IPortfolios {
    *
    * @param stocks the specified mapping of stocks and their quantity to be part of this portfolio.
    */
-  void createNewPortfolio(Map<String, Integer> stocks);
+  void createNewPortfolio(Map<String, Long> stocks);
 
   /**
    * Gets all the available portfolios.
